Remove base class and transplant its internals
