+++ app_store_scraper/app_store.py
@@ -1,4 +1,5 @@
 import logging
+import random
 import re
 import requests
 import sys
@@ -6,12 +7,29 @@ import time
 from datetime import datetime
 from requests.adapters import HTTPAdapter
 from requests.packages.urllib3.util.retry import Retry
-from .base import Base
 
 logger = logging.getLogger("AppStore")
 
 
-class AppStore(Base):
+class AppStore:
+    __scheme = "https"
+
+    __landing_host = "apps.apple.com"
+    __request_host = "amp-api.apps.apple.com"
+
+    __landing_path = "{country}/app/{app_name}/id{app_id}"
+    __request_path = "v1/catalog/{country}/apps/{app_id}/reviews"
+
+    __base_landing_url = f"{__scheme}://{__landing_host}"
+    __base_request_url = f"{__scheme}://{__request_host}"
+
+    __user_agents = [
+        # NOTE: grab from https://bit.ly/2zu0cmU
+        "Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1; SV1; .NET CLR 1.1.4322)",
+        "Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko",
+        "Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1)",
+    ]
+
     def __init__(
         self,
         country,
@@ -21,11 +39,38 @@ class AppStore(Base):
         log_level="INFO",
         log_interval=10,
     ):
-        super().__init__(country, app_name, app_id)
-        self.request_headers.update({"Authorization": self.__token()})
+        self.country = str(country).lower()
+        self.app_name = re.sub(r"[\W_]+", "-", str(app_name).lower())
+        self.app_id = str(app_id)
+
+        self.landing_url = self.__landing_url()
+        self.request_url = self.__request_url()
+
+        self.request_offset = 0
+        self.request_headers = {
+            "Accept": "application/json",
+            "Authorization": self.__token(),
+            "Connection": "keep-alive",
+            "Content-Type": "application/x-www-form-urlencoded; charset=UTF-8",
+            "Origin": self.__base_landing_url,
+            "Referer": self.landing_url,
+            "User-Agent": random.choice(self.__user_agents),
+        }
+        self.request_params = {
+            "l": "en-GB",
+            "offset": self.request_offset,
+            "limit": 20,
+            "platform": "web",
+            "additionalPlatforms": "appletv,ipad,iphone,mac",
+        }
+
+        self.reviews = list()
+        self.reviews_count = int()
+        self.fetched_count = int()
 
         logging.basicConfig(format=log_format, level=log_level.upper())
         self.log_interval = log_interval
+        self.log_timer = float()
 
     def __repr__(self):
         return "{object}(country={country}, app_name={app_name}, app_id={app_id})".format(
@@ -45,6 +90,16 @@ class AppStore(Base):
             f"{'Review count'.rjust(width, ' ')} | {self.reviews_count}"
         )
 
+    def __landing_url(self):
+        landing_url = f"{self.__base_landing_url}/{self.__landing_path}"
+        return landing_url.format(
+            country=self.country, app_name=self.app_name, app_id=self.app_id
+        )
+
+    def __request_url(self):
+        request_url = f"{self.__base_request_url}/{self.__request_path}"
+        return request_url.format(country=self.country, app_id=self.app_id)
+
     def __get(
         self,
         url,
@@ -60,7 +115,7 @@ class AppStore(Base):
             status_forcelist=status_forcelist,
         )
         with requests.Session() as s:
-            s.mount(self.base_request_url, HTTPAdapter(max_retries=retries))
+            s.mount(self.__base_request_url, HTTPAdapter(max_retries=retries))
             logger.debug(f"Making a GET request: {url}")
             self.response = s.get(url, headers=headers, params=params)
